+++ src/main/typescript/ruleEngine/messages/changedTypeMessage.ts
@@ -1,0 +1,29 @@
+import { getGame } from "../../foundryHelpers.js";
+import RuleElementMessage from "../ruleElementMessage.js";
+
+export default class ChangedTypeMessage extends RuleElementMessage {
+  constructor(
+    /** The name of the target document */
+    public docName: string | null,
+
+    /** The path to the property */
+    public propertyPath: string,
+
+    /** The name of the original type */
+    public originalType: string,
+
+    /** The name of the new type */
+    public newType: string
+  ) {
+    super("wv.ruleEngine.warnings.changedType", "warning");
+  }
+
+  override get message(): string {
+    return getGame().i18n.format(this.messageKey, {
+      name: this.docName,
+      path: this.propertyPath,
+      original: this.originalType,
+      new: this.newType
+    });
+  }
+}