+++ src/main/typescript/typesConfig.d.ts
@@ -14,7 +14,7 @@ import type {
 } from "./rolls/criticalsModifiers.js";
 import type * as settings from "./settings.js";
 import { macros } from "./macros/index.js";
-import type Ajv from "ajv/dist/jtd";
+import type Ajv from "ajv";
 
 declare global {
   interface SourceConfig {