+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- a value replacing rule element
+  ([#95](https://github.com/Wasteland-Ventures-Group/WV-VTT-module/issues/95))
+
 ### Changed
 
 - use esbuild to package the system